LPS-59258 Fix wrong argument datatype

+++ webs/screens-web/docroot/WEB-INF/src/com/liferay/screens/service/impl/ScreensJournalArticleServiceImpl.java
@@ -32,7 +32,7 @@ public class ScreensJournalArticleServiceImpl
 
 	@Override
 	public String getJournalArticleContent(
-			int groupId, int classPK, Locale locale)
+			long groupId, long classPK, Locale locale)
 		throws PortalException, SystemException {
 
 		JournalArticleResource journalArticleResource =
